+++ Modules/Intent.Modules.Modelers.Services.DomainInteractions/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.2
+
+- Improvement: CRUD Dialog no longer shows non-aggregate entities without primary keys.
+
 ### Version 1.0.0
 
 - New Feature: Services modeler extensions for describing interactions between an application's services and its domain.